refactor: split NewPostViewController Views into different UI Classes
